fix: extract components from page
