+++ pages/persoonsgegevens/[person].tsx
@@ -23,12 +23,12 @@ import Head from "next/head";
 import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import React, { ChangeEventHandler, FormEvent, useContext, useEffect, useState } from "react";
-import { AddressDataList } from "./sections/AddressDataList";
-import { PersonalDataList } from "./sections/PersonalDataList";
+import React, { ChangeEventHandler, FormEvent, useContext, useState } from "react";
 import { Aside, Checkbox2, OptionalIndicator, PageContentMain, ReservationCard } from "../../src/components";
+import { AddressDataList } from "../../src/components/huwelijksplanner/AddressDataList";
 import { PageFooterTemplate } from "../../src/components/huwelijksplanner/PageFooterTemplate";
 import { PageHeaderTemplate } from "../../src/components/huwelijksplanner/PageHeaderTemplate";
+import { PersonalDataList } from "../../src/components/huwelijksplanner/PersonalDataList";
 import { MarriageOptionsContext } from "../../src/context/MarriageOptionsContext";
 import { useIngeschrevenpersoonGetByBsn } from "../../src/hooks/useIngeschrevenpersoonGetByBsn";
 import { getBsnFromJWT } from "../../src/openapi/authentication";
